use as module, call function in api route
